Rationalizing module names and exported names
- The name of the module of config iterators is changed to
  `picire.iterator`, that of config splitters to `picire.splitter`,
  and that of outcome caches to `picire.cache`.
- Cache classes are not exported anymore in `picire`, but they
  should be accessed via the `picire.cache` module.
- `CombinedIterator` is moved next to `CombinedParallelDD` as that
  is the only algorithm that uses it.
