+++ picire/__init__.py
@@ -5,19 +5,19 @@
 # This file may not be copied, modified, or distributed except
 # according to those terms.
 
+from . import cache
 from . import cli
-from . import config_iterators
-from . import config_splitters
+from . import iterator
+from . import splitter
 from .abstract_dd import AbstractDD
 from .abstract_parallel_dd import AbstractParallelDD
+from .cache import CacheRegistry
 from .cli import __version__, reduce
-from .combined_iterator import CombinedIterator
-from .combined_parallel_dd import CombinedParallelDD
-from .config_iterators import IteratorRegistry
-from .config_splitters import SplitterRegistry
+from .combined_parallel_dd import CombinedIterator, CombinedParallelDD
 from .dd import DD
+from .iterator import IteratorRegistry
 from .outcome import Outcome
-from .outcome_cache import CacheRegistry, ConfigCache, ContentCache, OutcomeCache
 from .parallel_dd import ParallelDD
 from .shared_cache import shared_cache_decorator
+from .splitter import SplitterRegistry
 from .subprocess_test import ConcatTestBuilder, SubprocessTest
