+++ picire/iterator.py
