+++ tests/test_api.py
@@ -92,32 +92,32 @@ class TestApi:
         assert output == expect
 
     @pytest.mark.parametrize('split, subset_first, subset_iterator, complement_iterator, cache', [
-        (picire.config_splitters.BalancedSplit, True, picire.config_iterators.forward, picire.config_iterators.forward, picire.OutcomeCache),
-        (picire.config_splitters.ZellerSplit, True, picire.config_iterators.forward, picire.config_iterators.backward, picire.ConfigCache),
-        (picire.config_splitters.BalancedSplit, False, picire.config_iterators.backward, picire.config_iterators.forward, picire.OutcomeCache),
-        (picire.config_splitters.ZellerSplit, False, picire.config_iterators.backward, picire.config_iterators.backward, picire.ConfigCache),
-        (picire.config_splitters.BalancedSplit, True, picire.config_iterators.skip, picire.config_iterators.forward, picire.OutcomeCache),
-        (picire.config_splitters.ZellerSplit, True, picire.config_iterators.skip, picire.config_iterators.backward, picire.ConfigCache),
+        (picire.splitter.BalancedSplit, True, picire.iterator.forward, picire.iterator.forward, picire.cache.OutcomeCache),
+        (picire.splitter.ZellerSplit, True, picire.iterator.forward, picire.iterator.backward, picire.cache.ConfigCache),
+        (picire.splitter.BalancedSplit, False, picire.iterator.backward, picire.iterator.forward, picire.cache.OutcomeCache),
+        (picire.splitter.ZellerSplit, False, picire.iterator.backward, picire.iterator.backward, picire.cache.ConfigCache),
+        (picire.splitter.BalancedSplit, True, picire.iterator.skip, picire.iterator.forward, picire.cache.OutcomeCache),
+        (picire.splitter.ZellerSplit, True, picire.iterator.skip, picire.iterator.backward, picire.cache.ConfigCache),
     ])
     def test_dd(self, interesting, config, expect, granularity, split, subset_first, subset_iterator, complement_iterator, cache):
         self._run_picire(interesting, config, expect, granularity, picire.DD, split, subset_first, subset_iterator, complement_iterator, cache)
 
     @pytest.mark.parametrize('split, subset_first, subset_iterator, complement_iterator, cache', [
-        (picire.config_splitters.ZellerSplit, False, picire.config_iterators.forward, picire.config_iterators.forward, picire.ConfigCache),
-        (picire.config_splitters.BalancedSplit, False, picire.config_iterators.forward, picire.config_iterators.backward, picire.OutcomeCache),
-        (picire.config_splitters.ZellerSplit, True, picire.config_iterators.backward, picire.config_iterators.forward, picire.ConfigCache),
-        (picire.config_splitters.BalancedSplit, True, picire.config_iterators.backward, picire.config_iterators.backward, picire.OutcomeCache),
-        (picire.config_splitters.ZellerSplit, True, picire.config_iterators.skip, picire.config_iterators.forward, picire.ConfigCache),
-        (picire.config_splitters.BalancedSplit, True, picire.config_iterators.skip, picire.config_iterators.backward, picire.OutcomeCache),
+        (picire.splitter.ZellerSplit, False, picire.iterator.forward, picire.iterator.forward, picire.cache.ConfigCache),
+        (picire.splitter.BalancedSplit, False, picire.iterator.forward, picire.iterator.backward, picire.cache.OutcomeCache),
+        (picire.splitter.ZellerSplit, True, picire.iterator.backward, picire.iterator.forward, picire.cache.ConfigCache),
+        (picire.splitter.BalancedSplit, True, picire.iterator.backward, picire.iterator.backward, picire.cache.OutcomeCache),
+        (picire.splitter.ZellerSplit, True, picire.iterator.skip, picire.iterator.forward, picire.cache.ConfigCache),
+        (picire.splitter.BalancedSplit, True, picire.iterator.skip, picire.iterator.backward, picire.cache.OutcomeCache),
     ])
     def test_parallel(self, interesting, config, expect, granularity, split, subset_first, subset_iterator, complement_iterator, cache):
         self._run_picire(interesting, config, expect, granularity, picire.ParallelDD, split, subset_first, subset_iterator, complement_iterator, cache)
 
     @pytest.mark.parametrize('split, subset_first, subset_iterator, complement_iterator, cache', [
-        (picire.config_splitters.ZellerSplit, True, picire.config_iterators.forward, picire.config_iterators.forward, picire.OutcomeCache),
-        (picire.config_splitters.BalancedSplit, True, picire.config_iterators.forward, picire.config_iterators.backward, picire.ConfigCache),
-        (picire.config_splitters.ZellerSplit, False, picire.config_iterators.backward, picire.config_iterators.forward, picire.OutcomeCache),
-        (picire.config_splitters.BalancedSplit, False, picire.config_iterators.backward, picire.config_iterators.backward, picire.ConfigCache),
+        (picire.splitter.ZellerSplit, True, picire.iterator.forward, picire.iterator.forward, picire.cache.OutcomeCache),
+        (picire.splitter.BalancedSplit, True, picire.iterator.forward, picire.iterator.backward, picire.cache.ConfigCache),
+        (picire.splitter.ZellerSplit, False, picire.iterator.backward, picire.iterator.forward, picire.cache.OutcomeCache),
+        (picire.splitter.BalancedSplit, False, picire.iterator.backward, picire.iterator.backward, picire.cache.ConfigCache),
     ])
     def test_combined(self, interesting, config, expect, granularity, split, subset_first, subset_iterator, complement_iterator, cache):
         self._run_picire(interesting, config, expect, granularity, picire.CombinedParallelDD, split, subset_first, subset_iterator, complement_iterator, cache)
